refactoring : Modulize components
- Tab을 이루고 있는 2개의 components를 모듈화(TextTab, MapTab)
- App.js 코드 단순화
