+++ cleanNmap/TextTab.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import { Text } from 'react-native';
+
+const TextScreen = () => <Text>text</Text>;
+
+export default TextScreen;
